Split the tests in the ranger-client into stubs so that the testjar can be used across for various modules
