+++ ranger-client/src/test/java/com/flipkart/ranger/client/stubs/TestDeserializer.java
@@ -1,0 +1,6 @@
+package com.flipkart.ranger.client.stubs;
+
+import com.flipkart.ranger.core.model.Deserializer;
+
+public class TestDeserializer<T> implements Deserializer<T> {
+}
