+++ ranger-client/src/test/java/com/flipkart/ranger/client/stubs/TestShardInfo.java
@@ -1,0 +1,12 @@
+package com.flipkart.ranger.client.stubs;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class TestShardInfo {
+    private int shardId;
+}
