+++ ranger-http/src/test/java/com/flipkart/ranger/http/common/HttpNodeDataStoreConnectorTest.java
@@ -16,7 +16,6 @@
 package com.flipkart.ranger.http.common;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.flipkart.ranger.core.units.TestNodeData;
 import com.flipkart.ranger.http.config.HttpClientConfig;
 import org.junit.Assert;
 import org.junit.Test;
