+++ ranger-server/src/main/java/com/flipkart/ranger/server/AppConfiguration.java
@@ -4,7 +4,10 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.flipkart.ranger.client.RangerClientConstants;
 import com.flipkart.ranger.core.model.Service;
 import io.dropwizard.Configuration;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
