+++ ranger-server/src/main/java/com/flipkart/ranger/server/model/ShardInfo.java
@@ -17,7 +17,6 @@ package com.flipkart.ranger.server.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
-import lombok.extern.slf4j.Slf4j;
 
 import java.util.Objects;
