+++ ranger-zk-client/src/main/java/com/flipkart/ranger/client/zk/SimpleRangerZKClient.java
@@ -16,8 +16,8 @@
 package com.flipkart.ranger.client.zk;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.flipkart.ranger.client.RangerClientConstants;
 import com.flipkart.ranger.client.RangerClient;
+import com.flipkart.ranger.client.RangerClientConstants;
 import com.flipkart.ranger.core.finder.SimpleShardedServiceFinder;
 import com.flipkart.ranger.core.finder.shardselector.MatchingShardSelector;
 import com.flipkart.ranger.core.model.Criteria;
